Refactor element queries appliers & providers
