--- src/services/appliers/field/MatrixFieldContentApplier.php
@@ -1,15 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace lilthq\craftliltplugin\services\appliers\field;
-
-use lilthq\craftliltplugin\parameters\CraftliltpluginParameters;
-
-class MatrixFieldContentApplier extends AbstractElementQueryContentApplier implements ApplierInterface
-{
-    public function support(ApplyContentCommand $command): bool
-    {
-        return get_class($command->getField()) === CraftliltpluginParameters::CRAFT_FIELDS_MATRIX;
-    }
-}
